Refactored PatchEmbeddings into separate module.
