+++ tfimm/layers/__init__.py
@@ -1,4 +1,4 @@
 from .classifier import ClassifierHead  # noqa: F401
 from .drop import DropPath  # noqa: F401
 from .factory import act_layer_factory, norm_layer_factory  # noqa: F401
-from .transformers import MLP  # noqa:F401
+from .transformers import MLP, PatchEmbeddings  # noqa:F401
