+++ tfimm/layers/transformers.py
@@ -1,6 +1,49 @@
+"""
+Common layers shared between transformer architectures.
+
+Copyright 2021 Martins Bruveris
+"""
 import tensorflow as tf
 
-from tfimm.layers.factory import act_layer_factory
+from tfimm.layers.factory import act_layer_factory, norm_layer_factory
+
+
+class PatchEmbeddings(tf.keras.layers.Layer):
+    """
+    Image to Patch Embedding.
+    """
+    def __init__(
+        self,
+        patch_size: int,
+        embed_dim: int,
+        norm_layer: str = "",
+        **kwargs
+    ):
+        super().__init__(**kwargs)
+        self.patch_size = patch_size
+        self.embed_dim = embed_dim
+        self.norm_layer = norm_layer_factory(norm_layer)
+
+        self.projection = tf.keras.layers.Conv2D(
+            filters=self.embed_dim,
+            kernel_size=self.patch_size,
+            strides=self.patch_size,
+            use_bias=True,
+            name="proj",
+        )
+        self.norm = self.norm_layer(name="norm")
+
+    def call(self, x, training=False):
+        emb = self.projection(x)
+
+        # Change the 2D spatial dimensions to a single temporal dimension.
+        # shape = (batch_size, num_patches, out_channels=embed_dim)
+        batch_size, height, width = tf.unstack(tf.shape(x)[:3])
+        num_patches = (width // self.patch_size) * (height // self.patch_size)
+        emb = tf.reshape(tensor=emb, shape=(batch_size, num_patches, -1))
+
+        emb = self.norm(emb, training=training)
+        return emb
 
 
 class MLP(tf.keras.layers.Layer):
